+++ app/src/main/java/com/example/loomoapp/Loomo/LoomoRealSense.kt
@@ -120,5 +120,4 @@ class LoomoRealSense(private val publisher_: RealsensePublisher) {
             }
         }
     }
-
 }
