+++ app/src/main/java/com/example/loomoapp/OpenCV/OpenCVMain.kt
@@ -124,72 +124,7 @@ class OpenCVMain : Service() {
         }
     }
 
-    private fun ByteBuffer.toMat(width: Int, height: Int, cvType: Int): Mat {
-        val mat = Mat()
-        mat.create(height, width, cvType)
-        mat.put(0, 0, this.toByteArray())
-        return mat
-    }
 
-//    private fun writeToMain() {
-//        mHandler.post {
-//            mFishEyeByteBuffer.value = getFrame(fishEyeFrameBuffer.peek(0)!!)
-//            mColorByteBuffer.value = getFrame(colorFrameBuffer.peek(0)!!)
-//            mDepthByteBuffer.value = getFrame(depthFrameBuffer.peek(0)!!)
-//        }
-//    }
-
-    //    fun newFishEyeFrame(byteBuf: ByteBuffer) {
-//        newFrame(byteBuf, fishEyeFrame, FISHEYE_WIDTH, FISHEYE_HEIGHT, CV_8UC1)
-//        fishEyeFrame = fishEyeTracker.onNewFrame(fishEyeFrame)
-//    }
-//    fun newColorFrame(byteBuf: ByteBuffer) {
-//        newFrame(byteBuf, colorFrame, COLOR_WIDTH, COLOR_HEIGHT, CV_8UC4)
-//    }
-//    fun newDepthFrame(byteBuf: ByteBuffer) {
-//        newFrame(byteBuf, depthFrame, DEPTH_WIDTH, DEPTH_HEIGHT, CV_8UC2)
-//    }
-//
-//    fun getFishEyeFrame(): Bitmap {
-//        return getFrame(fishEyeFrame)
-//    }
-//    fun getColorFrame(): Bitmap {
-//        return getFrame(colorFrame)
-//    }
-//    fun getDepthFrame(): Bitmap {
-//        return getFrame(depthFrame)
-//    }
-//
-//    private fun newFrame(byteBuf: ByteBuffer, dst: Mat, width: Int, height: Int, cvType: Int) {
-//        dst.create(height, width, cvType)
-//        dst.put(0, 0, byteBuf.toByteArray())
-//    }
-//
-    private fun Mat.toBitmap(): Bitmap {
-        val bmp = Bitmap.createBitmap(this.cols(), this.rows(), Bitmap.Config.ARGB_8888)
-        // matToBitmap() only works for CV_8UC1, CV_8UC3 and CV_8UC4 formats,
-        // so the depth image's color space must be converted
-        if (this.type() == CV_8UC2) {
-            // The color space conversion for some reason does not change the frame.type(),
-            // so the workaround is make a new 'Mat' with a usable type
-            val tmp = Mat(DEPTH_HEIGHT, DEPTH_WIDTH, CV_8UC3)
-            cvtColor(this, tmp, COLOR_BGR5652RGB)
-            matToBitmap(tmp, bmp)
-            return bmp
-        }
-//        if (!(frame.type() == CV_8UC1 || frame.type() == CV_8UC3 || frame.type() == CV_8UC4)) {
-//        }
-        if (this.empty()) {
-            Log.d(TAG, "Frame is empty")
-            return bmp
-        }
-        try {
-            matToBitmap(this, bmp)
-        } catch (e: CvException) {
-            return bmp
-        }
-        return bmp
-    }
 
     fun checkFrame(streamType: Int, callback: (Bitmap) -> Unit){
         val frame: Mat? = when(streamType) {
